api--microseg-model/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
numpy==2.1.4
opencv-python==4.11.0.86
pip==23.2.1
protobuf==5.29.4
setuptools==78.1.0
